Refactor INFRAIFCIDSValidation: Split large files into smaller focused classes
- Created AvailableIdsFilesCollector to handle IDS file collection.
- Created AvailableProjectsCollector for project retrieval.
- Introduced AvailableValidationCommandsCollector for validation command options.
- Added IfcFileResolver for IFC file resolution logic.
- Implemented InfraApiCollectorHelpers for API interaction and logging.
- Added InfraApiWrapper to encapsulate API method calls.
- Defined InfraCommand enum for command types.
- Updated plan.md to outline refactoring steps and verification criteria.
