+++ src/Assistant/dotnet/INFRAIFCIDSValidation/InfraCommand.cs
@@ -1,0 +1,9 @@
+namespace INFRAIFCIDSValidation;
+
+public enum InfraCommand
+{
+    IFC_CHECK,
+    STEP_SYNTAX,
+    IFC_SCHEMA,
+    IDS_VALIDATION,
+}
